+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add Remote Access shortcut to settings sidebar
+
 ### Changed
 
 - Reconcile guides with v0.32.0 changes